Data.H3.Colour

Description

Scales that assign colours to values.

Assign values to colours

ordinalColours :: [a] -> ScaleOptions OrdinalColours a (Colour Double)

Create a scale that maps discrete values to colours. Colours will repeat if there are more than 12 different values.

data OrdinalColours a

Utilities and re-exports

data Colour a

This type represents the human preception of colour. The a parameter is a numeric type used internally for the representation.

The Monoid instance allows one to add colours, but beware that adding colours can take you out of gamut. Consider using blend whenever possible.

data PaletteType

Constructors

Sequential

Diverging

Qualitative

paletteFor :: PaletteType -> Int -> [Colour Double]

Get a palette of the specified size for a palette type. The actual size of the palette may be smaller or larger than the argument.

toCSSColour :: Colour Double -> String

Convert a colour to hexadecimal form so that it can be used in style attributes. Example: "#00aaff"